add another entity and mapper

+++ src/main/java/ru/kpfu/itis/kononenko/entity/Marriage.java
@@ -3,9 +3,9 @@ package ru.kpfu.itis.kononenko.entity;
 import java.sql.Date;
 
 public record Marriage(
-        int id,
-        int spouseMale,
-        int spouseFemale,
+        Long id,
+        Long spouseMale,
+        Long spouseFemale,
         Date startDate,
         Date endDate
 ) { }
