+++ src/main/java/ru/kpfu/itis/kononenko/entity/ParentChildRelation.java
@@ -1,7 +1,7 @@
 package ru.kpfu.itis.kononenko.entity;
 
 public record ParentChildRelation(
-        int id,
-        int parentId,
-        int childId
+        Long id,
+        Long parentId,
+        Long childId
 ) {}
